move some def + add jsDoc
